Extract NewsLink removals actions to a controller
