+++ src/views/news_link_removals/adding.phtml
